requirements.txt
black==25.1.0
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-WTF==1.2.1 
flit_core==3.12.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mypy_extensions==1.1.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.7
Werkzeug==3.1.3
flask_sqlalchemy
flask_login